Refactor Item class to inner class of Product
