+++ src/main/java/seedu/ibook/storage/JsonAdaptedProduct.java
@@ -9,13 +9,13 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import seedu.ibook.commons.exceptions.IllegalValueException;
-import seedu.ibook.model.item.Item;
 import seedu.ibook.model.item.UniqueItemList;
 import seedu.ibook.model.product.Category;
 import seedu.ibook.model.product.Description;
 import seedu.ibook.model.product.Name;
 import seedu.ibook.model.product.Price;
 import seedu.ibook.model.product.Product;
+import seedu.ibook.model.product.Product.Item;
 
 /**
  * Jackson-friendly version of {@link Product}.
